pyproject.toml
[tool.poetry]
name = "drf-course"
version = "0.1.0"
description = ""
authors = ["bbt-t <zlukcss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django = "^4.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
